requirements.txt
# Application requirements
Werkzeug==0.16.0
flask==1.0.2
flask-restplus==0.12.1
flask-sqlalchemy==2.3.1
flask-migrate
pytest-flask==0.14.0
freezegun==0.1.11
faker==1.0.7
parse==1.12.0
requests==2.23.0
flask-cors
flask-marshmallow